Made changes to the result of the Fourier Transform. FrequencyFilter now contains the bins in the mirrored half of the fourier series. Waves are now recreated and rendered correctly.

+++ AudioVisual/Utils/MathUtils.cs
@@ -105,7 +105,7 @@ namespace AudioVisual.Utils
             return result;
         }
 
-        public static List<double> NormalizeValues(IReadOnlyList<float> values)
+        public static List<double> NormalizeValues(IReadOnlyList<double> values)
         {
             var maxAmplitude = values.Max(Math.Abs);
             const double epsilon = 1e-3;
